most recent commit has a frontend photo uploader, new model and schema, and a model and schema for images
