--- client/index.html
@@ -1,39 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-
-    <title>Bike Trader</title>
-
-</head>
-<body>
-
-	<body ng-app="fileUpload" ng-controller="MyCtrl">
-  <h4>Upload on file select</h4>
-  <button type="file" ngf-select="uploadFiles($file, $invalidFiles)"
-          accept="image/*" ngf-max-height="1000" ngf-max-size="1MB">
-      Select File</button>
-  <br><br>
-  File:
-  <div style="font:smaller">{{f.name}} {{errFile.name}} {{errFile.$error}} {{errFile.$errorParam}}
-      <span class="progress" ng-show="f.progress >= 0">
-          <div style="width:{{f.progress}}%"  
-               ng-bind="f.progress + '%'"></div>
-      </span>
-  </div>     
-  {{errorMsg}}
-</body>
-
-
-
-  trading bikes!
-
-  <!-- scripts etc -->
-	<script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.5/angular.min.js"></script>
-	<!-- 1. add ng file upload via npm install, 2. add a script tag for it here -->
-	<script src="../node_modules/ng-file-upload/dist/ng-file-upload.min.js"></script>
-
-	<!-- custom angular -->
-	<script src="app.js"></script>
-</body>
-</html>
