+++ public/index.html
@@ -10,6 +10,7 @@
 
   trading bikes!
   <h1>hello from the index page</h1>
+
   <div ng-view></div>
   <p> {{ tagline }} </p>
